titan-voice-server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
faster-whisper>=1.1.0
hume-tada>=0.1.0
torch>=2.5.0
torchaudio>=2.5.0
httpx>=0.28.0
numpy>=1.26.0
